+++ src/main/kotlin/com/back/koreaTravelGuide/domain/ai/weather/dto/LandForecastData.kt
@@ -1,0 +1,3 @@
+package com.back.koreaTravelGuide.domain.ai.weather.dto
+
+data class LandForecastData()
