+++ src/main/kotlin/com/back/koreaTravelGuide/domain/ai/weather/dto/TemperatureData.kt
@@ -4,7 +4,7 @@ import kotlin.text.set
 
 // TODO: 날씨 내부 데이터 구조 - 기상청 API 응답 데이터 매핑용 내부 클래스들
 @Suppress("unused") // JSON 직렬화를 위해 필요
-data class TemperatureDto(
+data class TemperatureData(
     private val days: MutableMap<Int, TemperatureInfo?> = mutableMapOf(),
 ) {
     fun setDay(
