+++ src/main/kotlin/com/back/koreaTravelGuide/domain/ai/weather/service/tools/Tools.kt
@@ -1,4 +1,4 @@
-package com.back.koreaTravelGuide.domain.weather.service.tools
+package com.back.koreaTravelGuide.domain.ai.weather.service.tools
 
 import org.springframework.stereotype.Component
 import java.time.ZoneId
@@ -51,15 +51,52 @@ open class Tools {
     }
 
     companion object {
-        private val REGION_MAP =
-            mapOf(
-                "서울" to "11B10101", "인천" to "11B20201", "수원" to "11B20601", "파주" to "11B20305",
-                "이천" to "11B20612", "평택" to "11B20606", "춘천" to "11D10301", "원주" to "11D10401",
-                "강릉" to "11D20501", "속초" to "11D20601", "대전" to "11C20401", "세종" to "11C20404",
-                "청주" to "11C10301", "충주" to "11C10101", "전주" to "11F10201", "군산" to "11F10501",
-                "광주" to "11F20501", "목포" to "11F20401", "여수" to "11F20801", "대구" to "11H10701",
-                "안동" to "11H10501", "포항" to "11H10201", "부산" to "11H20201", "울산" to "11H20101",
-                "창원" to "11H20301", "통영" to "11H20401", "제주" to "11G00201", "서귀포" to "11G00401",
-            )
+        private val REGION_MAP = mapOf(
+            "백령도" to "11A00101", "서울" to "11B10101", "과천" to "11B10102", "광명" to "11B10103", "강화" to "11B20101",
+            "김포" to "11B20102", "인천" to "11B20201", "시흥" to "11B20202", "안산" to "11B20203", "부천" to "11B20204",
+            "의정부" to "11B20301", "고양" to "11B20302", "양주" to "11B20304", "파주" to "11B20305", "동두천" to "11B20401",
+            "연천" to "11B20402", "포천" to "11B20403", "가평" to "11B20404", "구리" to "11B20501", "남양주" to "11B20502",
+            "양평" to "11B20503", "하남" to "11B20504", "수원" to "11B20601", "안양" to "11B20602", "오산" to "11B20603",
+            "화성" to "11B20604", "성남" to "11B20605", "평택" to "11B20606", "의왕" to "11B20609", "군포" to "11B20610",
+            "안성" to "11B20611", "용인" to "11B20612", "이천" to "11B20701", "광주" to "11B20702", "여주" to "11B20703",
+            "충주" to "11C10101", "진천" to "11C10102", "음성" to "11C10103", "제천" to "11C10201", "단양" to "11C10202",
+            "청주" to "11C10301", "보은" to "11C10302", "괴산" to "11C10303", "증평" to "11C10304", "추풍령" to "11C10401",
+            "영동" to "11C10402", "옥천" to "11C10403", "서산" to "11C20101", "태안" to "11C20102", "당진" to "11C20103",
+            "홍성" to "11C20104", "보령" to "11C20201", "서천" to "11C20202", "천안" to "11C20301", "아산" to "11C20302",
+            "예산" to "11C20303", "대전" to "11C20401", "공주" to "11C20402", "계룡" to "11C20403", "세종" to "11C20404",
+            "부여" to "11C20501", "청양" to "11C20502", "금산" to "11C20601", "논산" to "11C20602", "철원" to "11D10101",
+            "화천" to "11D10102", "인제" to "11D10201", "양구" to "11D10202", "춘천" to "11D10301", "홍천" to "11D10302",
+            "원주" to "11D10401", "횡성" to "11D10402", "영월" to "11D10501", "정선" to "11D10502", "평창" to "11D10503",
+            "대관령" to "11D20201", "태백" to "11D20301", "속초" to "11D20401", "고성" to "11D20402", "양양" to "11D20403",
+            "강릉" to "11D20501", "동해" to "11D20601", "삼척" to "11D20602", "울릉도" to "11E00101", "독도" to "11E00102",
+            "전주" to "11F10201", "익산" to "11F10202", "정읍" to "11F10203", "완주" to "11F10204", "장수" to "11F10301",
+            "무주" to "11F10302", "진안" to "11F10303", "남원" to "11F10401", "임실" to "11F10402", "순창" to "11F10403",
+            "군산" to "21F10501", "김제" to "21F10502", "고창" to "21F10601", "부안" to "21F10602", "함평" to "21F20101",
+            "영광" to "21F20102", "진도" to "21F20201", "완도" to "11F20301", "해남" to "11F20302", "강진" to "11F20303",
+            "장흥" to "11F20304", "여수" to "11F20401", "광양" to "11F20402", "고흥" to "11F20403", "보성" to "11F20404",
+            "순천시" to "11F20405", "광주" to "11F20501", "장성" to "11F20502", "나주" to "11F20503", "담양" to "11F20504",
+            "화순" to "11F20505", "구례" to "11F20601", "곡성" to "11F20602", "순천" to "11F20603", "흑산도" to "11F20701",
+            "목포" to "21F20801", "영암" to "21F20802", "신안" to "21F20803", "무안" to "21F20804", "성산" to "11G00101",
+            "제주" to "11G00201", "성판악" to "11G00302", "서귀포" to "11G00401", "고산" to "11G00501", "이어도" to "11G00601",
+            "추자도" to "11G00800", "울진" to "11H10101", "영덕" to "11H10102", "포항" to "11H10201", "경주" to "11H10202",
+            "문경" to "11H10301", "상주" to "11H10302", "예천" to "11H10303", "영주" to "11H10401", "봉화" to "11H10402",
+            "영양" to "11H10403", "안동" to "11H10501", "의성" to "11H10502", "청송" to "11H10503", "김천" to "11H10601",
+            "구미" to "11H10602", "군위" to "11H10707", "고령" to "11H10604", "성주" to "11H10605", "대구" to "11H10701",
+            "영천" to "11H10702", "경산" to "11H10703", "청도" to "11H10704", "칠곡" to "11H10705", "울산" to "11H20101",
+            "양산" to "11H20102", "부산" to "11H20201", "창원" to "11H20301", "김해" to "11H20304", "통영" to "11H20401",
+            "사천" to "11H20402", "거제" to "11H20403", "고성" to "11H20404", "남해" to "11H20405", "함양" to "11H20501",
+            "거창" to "11H20502", "합천" to "11H20503", "밀양" to "11H20601", "의령" to "11H20602", "함안" to "11H20603",
+            "창녕" to "11H20604", "진주" to "11H20701", "산청" to "11H20703", "하동" to "11H20704", "사리원" to "11I10001",
+            "신계" to "11I10002", "해주" to "11I20001", "개성" to "11I20002", "장연(용연)" to "11I20003", "신의주" to "11J10001",
+            "삭주(수풍)" to "11J10002", "구성" to "11J10003", "자성(중강)" to "11J10004", "강계" to "11J10005", "희천" to "11J10006",
+            "평양" to "11J20001", "진남포(남포)" to "11J20002", "안주" to "11J20004", "양덕" to "11J20005", "청진" to "11K10001",
+            "웅기(선봉)" to "11K10002", "성진(김책)" to "11K10003", "무산(삼지연)" to "11K10004", "함흥" to "11K20001", "장진" to "11K20002",
+            "북청(신포)" to "11K20003", "혜산" to "11K20004", "풍산" to "11K20005", "원산" to "11L10001", "고성(장전)" to "11L10002",
+            "평강" to "11L10003",
+        )
     }
+
+
+
+
 }
